fixed up all the tests and added some
