--- tests/regression/26-undefined_behavior/09-struct-nullpointer.c
@@ -1,15 +1,0 @@
-// PARAM: --enable ana.nullptr --enable dbg.debug
-#include <stdio.h>
-struct person {
-  int age;
-  int height;
-};
-
-int main() {
-  struct person *personPtr;
-  personPtr = NULL;
-
-  personPtr->age = 24;
-
-  return 0;
-}
