+++ tests/regression/26-undefined_behavior/10-nullpointer-dereference-simple.c
@@ -1,0 +1,12 @@
+// PARAM: --enable ana.nullptr --enable dbg.debug
+#include <stdio.h>
+int main(){    
+    int *pi;     // a pointer to an integer
+    int *t;
+    int a = *t; // WARN
+    t = &pi;
+    pi = NULL; 
+    int c = *pi; // WARN
+    int d = *t; // NOWARN
+    return 1;
+}
