--- tests/regression/26-undefined_behavior/10-nullpointer-dereference.c
@@ -1,12 +1,0 @@
-// PARAM: --enable ana.nullptr --enable dbg.debug
-#include <stdio.h>
-int main() {
-  int *pi; // a pointer to an integer
-  int *t;
-  int a = *t; // WARN
-  t = &pi;
-  pi = NULL;
-  int c = *pi; // WARN
-  int d = *t;  // NOWARN
-  return 1;
-}
