+++ tests/regression/26-undefined_behavior/11-nullpointer-dereference-ints.c
@@ -4,20 +4,16 @@
 // source base:
 // https://stackoverflow.com/questions/4007268/what-exactly-is-meant-by-de-referencing-a-null-pointer
 int main() {
-  int arr[] = {1, 2, 3, 4, 5, 6};
   int a, b, c, x; // some integers
   int *pi;        // a pointer to an integer
   int *ok;
-  a = arr[2];
+  a = 5;
   ok = &a;
   pi = &a; // pi points to a
   b = *pi; // b is now 5
   pi = NULL;
-  c = *ok; // prolly nowarn
-  c = *pi; // prolly warn
-  arr[3] = NULL;
-  x = arr[3];
-  c = x; // prolly warn
+  x = *ok; //NOWARN
+  c = *pi; //WARN
 
   return 1;
 }
