+++ tests/regression/26-undefined_behavior/12-nullpointer-dereference-array.c
@@ -4,16 +4,20 @@
 // source base:
 // https://stackoverflow.com/questions/4007268/what-exactly-is-meant-by-de-referencing-a-null-pointer
 int main() {
+  int*  arr[] = {NULL, NULL, NULL, NULL, NULL, NULL};
   int a, b, c, x; // some integers
   int *pi;        // a pointer to an integer
   int *ok;
-  a = 5;
+  a = arr[2];
   ok = &a;
   pi = &a; // pi points to a
   b = *pi; // b is now 5
   pi = NULL;
-  x = *ok; // Prolly nowarn
-  c = *pi; // prolly warn
+  c = *ok; // NOWARN
+  c = *pi; //WARN
+  int t = *arr[3];//WARN
+  int* arr_2d[2][2] = {NULL, NULL, NULL, NULL};
+  int m = *arr_2d[1][1]; //WARN
 
   return 1;
 }
